Add memory-efficient training for consumer GPUs (24-32GB VRAM)
Features:
- RamTorch CPU offloading to train 19B model on consumer hardware
- Comprehensive TRAINING_GUIDE.md with step-by-step instructions
- Tested configurations for RTX 4090 (32GB) with ~7s/step
- HQQ and Quanto quantization support (though BF16+ramtorch is faster)
- Updated .gitignore for data directories

Key findings:
- BF16 + 60% ramtorch offload: ~7s/step (recommended)
- INT8-quanto + ramtorch: ~8s/step (slower due to dequant overhead)
- FP8-quanto + ramtorch: ~10s/step (even slower)
- FP8 native model: doesn't work (Float8_e4m3fn has no autograd)
- HQQ INT4: doesn't work (INT tensors can't have gradients)

Validation disabled during training (OOMs with ramtorch) - test checkpoints manually.

+++ packages/ltx-core/src/ltx_core/model/transformer/model.py
@@ -329,9 +329,6 @@ class LTXModel(torch.nn.Module):
         # Process transformer blocks
         for block in self.transformer_blocks:
             if self._enable_gradient_checkpointing and self.training:
-                # Use gradient checkpointing to save memory during training.
-                # With use_reentrant=False, we can pass dataclasses directly -
-                # PyTorch will track all tensor leaves in the computation graph.
                 video, audio = torch.utils.checkpoint.checkpoint(
                     block,
                     video,
@@ -358,8 +355,10 @@ class LTXModel(torch.nn.Module):
     ) -> torch.Tensor:
         """Process output for LTXV."""
         # Apply scale-shift modulation
+        # Cast both tensors to x.dtype to prevent float32 contamination with FP8 quantization
         scale_shift_values = (
-            scale_shift_table[None, None].to(device=x.device, dtype=x.dtype) + embedded_timestep[:, :, None]
+            scale_shift_table[None, None].to(device=x.device, dtype=x.dtype)
+            + embedded_timestep[:, :, None].to(dtype=x.dtype)
         )
         shift, scale = scale_shift_values[:, :, 0], scale_shift_values[:, :, 1]
